+++ app/lib/navigation-rules/hospital-inpatient.js
@@ -5,7 +5,7 @@ const appLogger = Logger();
 const hospitalInpatient = (req) => {
   if (req.journeyData.getDataForPage('hospital-inpatient').hospitalInpatient === 'no') {
     appLogger.info('Navigation rules: hospital-inpatient - not inpatient; make sure hospital details cleared from journey');
-    genericDataUtils.deleteIfPresent(req, 'hospital-inpatient', ['hospitalName', 'hospitalWard', 'admissionDate']);
+    genericDataUtils.deleteIfPresent(req, 'hospital-details', ['hospitalName', 'hospitalWard', 'admissionDate']);
   }
 };