+++ test/unit/lib/DataMapper.spec.js
@@ -96,13 +96,15 @@ describe('DataMapper', () => {
     },
     address: {
       correspondence: 'no',
-      correspondenceAddress: {
+      address: {
         address1: 'test',
         address2: 'test',
         address3: 'test',
         postcode: 'test',
       },
-      address: {
+    },
+    'correspondence-address': {
+      correspondenceAddress: {
         address1: 'test',
         address2: 'test',
         address3: 'test',
@@ -130,6 +132,8 @@ describe('DataMapper', () => {
     },
     'hospital-inpatient': {
       hospitalInpatient: 'yes',
+    },
+    'hospital-details': {
       admissionDate: {
         dd: '01',
         mm: '01',
@@ -200,6 +204,8 @@ describe('DataMapper', () => {
     },
     pregnant: {
       pregnant: 'yes',
+    },
+    'pregnant-due-date': {
       dueDate: {
         dd: '01',
         mm: '01',